cc+vm: finalize calling convention and InnerNode::{Call, Fn} handling
This commit finalizes the calling convention thusfar, each argument to a
function has to be in r0..rN, where N is the amount of arguments to said
function. For an example, see the following trace:

    (fn sum_three (a b c)
      (Plus
        a
        (Plus
          b
          c
        )
      )
    )
    (sum_three
      Integer("1")
      Integer("2")
      Integer("3")
    )
    __entry:
            jmp 5
    __sum_three:
    ; 0x0001 args=3;size=4
            add r3, r1, r2
            add r4, r0, r3
            mov r0, r4
            ret
            load_imm r0, \#1
            load_imm r1, \#2
            load_imm r2, \#3
            call sum_three

This is peephole optimised, but with the non-optimised version, which
includes self moves in this trivial example, one can see the calling
convention layout being prepared:

            ; load immediate for argument 0
            load_imm r0, \#1
            ; move argument 0 value into r0
            mov r0, r0

            ; load immediate for argument 1
            load_imm r1, \#2
            ; move argument 0 value into r1
            mov r1, r1

            ; load immediate for argument 2
            load_imm r2, \#3
            ; move argument 0 value into r2
            mov r2, r2

            call sum_three

This change enables the removal of `args_start` and `args_len` from both
the `Sys` and `Call` op codes.

+++ src/cc/ctx.rs
@@ -29,7 +29,7 @@ impl<'cc> Local<'cc> {
 #[derive(Debug, Clone)]
 pub struct Func<'cc> {
     pub name: &'cc str,
-    pub args: usize,
+    pub args: u8,
     pub size: usize,
     pub pc: usize,
 }
